Promote scan/canonical/check/extract/model to pkg/ for external consumers
The sibling audiosilo-sidecars tool needs the folder scanner, canonical
JSON writer, data validator, and ngram checker as ordinary module deps
(precedent: audiosilo-server's pkg/launcher + pkg/match for the manager).

- internal/{scan,canonical,check,extract,model} -> pkg/* (model comes
  along because pkg/check exports Result.Catalog *model.Catalog)
- pkg/scan keeps its within-module import of internal/importer's
  normalization helpers; nothing internal leaks into exported APIs
- add fenced ```json inline-attachment tests for the issueform sidecar
  path (already implemented, previously untested variant)
- update CLAUDE.md package layout + path references
