+++ pkg/model/location_test.go
