+++ pkg/model/model.go
@@ -1,6 +1,10 @@
 // Package model defines the on-disk entity types and the rules that tie a file
 // to its slug, shard, and location. It has no dependency on validation or
 // storage; both metacheck and metabuild build on it.
+//
+// This package is PUBLIC API: it is consumed by the sibling audiosilo-sidecars
+// tool as an ordinary module dependency (and reached through pkg/check's
+// exported Catalog), so its exported surface is a contract.
 package model
 
 // Kind identifies an entity type.
