+++ pkg/scan/derive.go
