+++ pkg/scan/scan_test.go
