+++ pkg/scan/split.go
