+++ pkg/scan/split_test.go
