+++ pkg/scan/tags.go
