+++ pkg/scan/tags_test.go
